Move Saswell TRV from moes.js to it's own file (#2593)
